requirement.txt
langchain
chromadb
faiss-cpu
sentence-transformers
pillow
opencv-python
streamlit
ollama
python-dotenv
pdfplumber
langchain_experimental
langchain-community
langchain-text-splitters
sentence-transformers
langchain-huggingface
streamlit
tabulate